PolicyPulse – Intelligent HR Policy Assistant/requirements.txt
langchain
langchain-community
langchain-qdrant
sentence-transformers
qdrant-client
pypdf
fastapi
uvicorn
python-multipart
python-dotenv
groq
streamlit
requests
